requirements.txt
streamlit
pandas
easyocr
openpyxl
opencv-python-headless
pillow